pyproject.toml
[project]
name = "photo-grouping"
version = "0.1.0"
description = "Photo grouping application using face detection and clustering"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.4",
    "huggingface-hub>=1.1.4",
    "numpy>=2.3.4",
    "onnxruntime>=1.23.2",
    "pillow>=12.0.0",
    "psutil>=6.1.0",
    "scikit-learn>=1.7.2",
    "ultralytics>=8.3.228",
]
